Add container sharding to the core of swift
To activate it pass:

  X-Container-Sharding: yes

Once activated, new metadata is added to the container (which
will now be called the root container):

  - shard_power
  - shard_bitmap
  - max_shard_objects

shard_power:

This works almost identically to part_power, and is uses the same
path hashing and part_power mechanisms other parts of swift uses.
But is a list (eventually only containing 1 value). (See later).

shard_bitmap:

This is a bitmap of the shard containers that exist. This is used
by the container reconciler when migrating up and down a shard_power.
(See later).

Shard containers are only created when an object is added to the
root container, and when the shard container doesn't exist.

max_shard_objects:

This is a rough average of objects per shard container, which can be
set by the user.

This is what swift uses to determine whether the shard_power needs to
increase:

  avg_objects_per_shard = num_objects / 2^shard_power
  if avg_objects_per_shard > max_shard_objects:
      shard_power.append(shard_power[-1]++)

TLDR:

Once activated, when an object is then put into the container, the hash
is calculated and the shard-power is used to generate a 'shard'
number. Which determines which container shard to store this object
in. If the container shard doesn't exist it is created.

                                      +---------------------+
        +-----------------+           |  Acct/cont_shard_0  |
        |    Acct/cont    |           +---------------------+
        +-----------------+           |Objs:                |
        |Metadata:        |           |   Obj3              |
        |  shard-power: 2 +---------> |   Obj1              |
        |  shards: 0,1    |           +---------------------+
        |                 |
        |Objs:            |           +---------------------+
        |                 |           |  Acct/cont_shard_1  |
        |                 |           +---------------------+
        |                 +---------> |Objs:                |
        |                 |           |   Obj2              |
        +-----------------+           |   Obj4              |
                                      +---------------------+

When a request is issued to the shard container (/Acct/cont above)
directly, the list of shards is iterated and it is run on all shard
containers. When its a container listing the bodies from all
sub-requests are returned.

This leads to a downside to sharding, the container listings returned
will not
be properly sorted, the result of each sub request will break when they
are all
combined.

Taking the diagram above as an example:

  curl -i -X GET -H "X-Auth-Token: $TOKEN" $STORAGE_URL/Acct/cont

  Obj1
  Obj3
  Obj2
  Obj4

Container sharding (More detail):

The shard_power is changed automatically by swift, so the amount of
sharding will grow with the amount of objects. To handle this the
shard_power metadata value isn't simply an int, it is a list of ints.

  [2, 3, ..]

Where the leftmost integer is the oldest shard_power, the right most
the newest:

  [<old-shard-power>, <newer-shard-power>, <newest-shard-power>]

When an object is added the right most shard_power is used for
placement.

What actually  happens is on a object PUT is the number of objects
(total) in the container is checked, and if max_shard_objects has
been reached, the (current shard power +1) is appended to the
shard_power list and the shard_bitmap is extended.

On an object GET, swift will attempt to find it using the right most
shard_power, if that fails will walk backward down the list until it
is found or if it doesn't exist a 404 will be returned.

When the length of the shard_list is greater then 1, this causes extra
overhead, but this allows the shard_power to grow. The container
reconciler (or maybe a new service) in the background, once has found a
sharded root container, will work through the list from left to right
and move objects into the correct sharded container. Once it has
completed a pass or the leftmost list entry it will be removed,
allowing sharding to be eventually consistent to 1 shard_power.

On and object put, if the required sharded_container doesn't exist, the
container will be created. Once done the shard_bitmap will be updated.
This shard_bitmap is what the container reconciler (or other service)
will
use to know what shard containers to hit.

The shard_bitmap is hexstring representing what shards have been
created.
A new class called Bitmap has been added to swift/common/utils to handle
the bitmap. The container reconciler, will need to shrink the bitmap
down once
sharded containers larger then the current shard_power have been moved.
All bitmap maintenance is handled via the class.

The Bitmap class has an inter() method to work its way through existing
shard
numbers:

  >> from swift.common.utils import Bitmap
  >> bitmap = Bitmap(6)
  >> print(bitmap)
  0000000000000000

  >> for x in (2, 4, 8, 15, 16, 17, 18, 32, 40):
  >>     bitmap.set_bit(x)

  >> print(bitmap)
  2881e00080800000

  >> bitmap.unset_bit(16)

  >> for x in bitmap.iter():
  >>     print x
  2
  4
  8
  15
  17
  18
  32
  40

All container metadata is stored in the root container, meaning all
sharded
containers have minimal inforamtion. In fact the only extra peice of
metadata
they have is:

  shard_container: <root container name>

That way if a shard container gets lost somehow (is missed in the bitmap
or failed
delete) we can itentify it's owner. (Although the name should give it
away, so maybe
this isn't required).

Change-Id: Ibad6c43f2cf9eb99188c3edb05c07f5db1106db1