[merge] Added basic shardtrie compression
This adds a ShardTrie.dump_to_zlib and load_from_zlib which
wraps and compresses dump_to_json and load_from_json methods.

Added a utils.shard_to_string helper method so there is once place
we call to dump a shardtrie to a string. Incase we need to do
something more funky in the future.
For example, after debugging change it to dump_to_zlib and pick
the level of compression.

Change-Id: Ibc50a56128f96600619d0f67698bfa556f9147d0