[merge] Replicator and db_replicator hackery
This change is attempt #1 at getting the replicator to replicate
node items from the shard_nodes table.

Because most of the container replicator code is apart of the
db_replicator parent class. This change introduces a hook to get
more objects to sync. Container backend merge_items is already
node item aware.

Change-Id: Id8577e26f5179311c41b78cee25a4708b4003019