[Merge] Added new CountingTrie class
This trie is a prefix trie, but is solely used to find the best
candidate sub tries for splitting. The trie doesn't actually store the
leaf nodes, it only increments counters on all nodes in the path of
placing the object.

If a nodes count ever becomes > the max_group_size, then it calls
the new_candidate method which if it is of a higher level the the
current best candidate then will keep track of it.

To keep this memory effiecnt, when attempting to place an object, if
there are other children in any node that the current object is not using,
they are deleted. The means you must add objects to this trie in the
correct order. Which we do thanks to always building from a database.

Change-Id: Iab12ae4f56ce34c18ae6b1b57056470481cee56f