[Merge] cont controllers GETorHEAD_sharded dist_node fix
GETorHEAD_sharded in the proxy container controller wasn't adding
visted dist_nodes to the dist_node list. With the new changes
where there is a copy of every dist_node in the root container, in
order to short circuit PUTs, meant a GET was visiting dist_node more
then once.

Change-Id: I8cccdb51e12a570c90bd8afdf8f5304fd0d8f7ce