Sped up the container_info and get_container_info
On every HEAD request to a sharded container, extra requests
are called to return the correct bytes_used and object_count,
this is awesome, but means there is overhead to every call to
container_info and get_container_info (for a sharded container).

This change adds the ability to pass skip_sharded to container_info
and get_container_info methods. If True, the root container is the
only container queried, meaning all metadata is a correct, but
bytes_used and object_count will be wrong.

get_container_info is mainly used by middleware, so skip_sharded is
defaulted to False.. so writers can trust the data coming back.
Continer_info on the other hand is only used by proxy controllers,
e.g. on every obj PUT, so to speed this up, it's defaulted to True.
(At least for now).

I have yet to debug this, will do that now.

Change-Id: I2f7792789965a151abefa98b76dd307e523dce8c