Use less reqests on 301 container updates
The first version of the 301 redirect following added to sharding
would generate the required headers for all primary nodes to update
on each request, creating a kind of request storm.

This change only adds 1 new node to update on a 301. It picks the same
primary node at the same index of the current container being updated.

Change-Id: I2b9ea46cca872c5608d4672b8ba1f8b8aa2ff63d